Refacto : Move tempo service outside domain
